+++ src/patches/chaos_drone.rs
@@ -97,11 +97,12 @@ different due to the chaotic nature of the algorithm.
     patch.add_connection("flt-1", "out", "amp-1", "in");
     patch.add_connection("env-1", "out", "amp-1", "cv");
     patch.add_connection("lfo-1", "out", "mth-1", "param_a");
-    patch.add_connection("amp-1", "out", "rev-1", "in_l");
-    patch.add_connection("rev-1", "out_l", "scp-1", "in_l");
-    patch.add_connection("rev-1", "out_r", "scp-1", "in_r");
-    patch.add_connection("scp-1", "out_l", "out-1", "in_l");
-    patch.add_connection("scp-1", "out_r", "out-1", "in_r");
+    // Voice output: amp -> stereo output (effects handled via effect chain)
+    patch.add_connection("amp-1", "left", "out-1", "in_l");
+    patch.add_connection("amp-1", "right", "out-1", "in_r");
+    // Oscilloscope taps from amplifier for visualization
+    patch.add_connection("amp-1", "left", "scp-1", "in_l");
+    patch.add_connection("amp-1", "right", "scp-1", "in_r");
 
     patch.settings.octave_offset = -1;
     patch